Rework rx fifo
Having two interrupt lines is not very useful if you have to synchronize
access to the monolithic Can type to do anything about the interrupt. If
the type could be destructured into smaller components, the need for
synchronization between the two interrupts would be reduced.

The logic is mostly the same as the old CanReadFifo, but without
duplicating the reads of the status register.

Since RxFifo borrows a part of SharedMemory, the other parts are
borrowed to a placeholder struct to allow this to compile until those
parts can be similarly refactored.