Enable bit timing configuration for bit-rate switched data (#3)
TimingParams is split to exploit the similarities between the bit timing
configuration in NBTP and DBTP. Timestamping is just moved to a separate
struct for now.

The validity of the parameters in the new BitTiming struct now checked
when applying them instead of on construction. This prevents public
fields in the struct allowing the user to skip the checks.

The divider (now called `prescaler` to better match the manual)
calculation is changed to remove floating point operations.