Split `mcan` into two crates + `mcan-core` documentation (#6)
In order to simplify the integration process of `mcan` into target HALs,
it was split into a main crate (`mcan`) and a small integration layer crate
(`mcan-core`) that contains only traits relevant on the HAL level. Such a
split allows for easier maintenance and a faster pace of development as
`mcan` releases will be less dependent on the release pace of target
HALs.

This commit also provides complete documentation of `mcan-core`.