Add method to check if bus is dominant (#51)
Checking if the bus is dominant (i.e., the RX pin value is low) can be
particularly useful in scenarios where the transceiver is put into sleep
mode. Some transceivers (such as TCAN1043-Q1[1]) signals wake-up by
setting the RX pin (and nFAULT) to low while being in sleep mode. To
accurately tell if the transceiver requests a wake-up, the RX pin state
needs to be accessible from the application.

Fortunately MCAN exposes a way to read the state of the RX pin through
its `TEST` register. Extend the `DynAux` trait with an additional method
that reads the `RX` bit from that register.

[1]: https://www.ti.com/product/TCAN1043-Q1